Added Logging and Lockfile Capability
Added a lockfile to work around and issue discovered by user wrender, where Plex deletes a grab folder once any script completes.  This means that a if a second or third script is running, it will fail to move the file back to the grab folder.  In this update, we create a lock file and all of the scripts will stall until the lock files are all deleted.  

This update also adds more logging, which is dumped into the /tmp folder for each postprocessing session.  Mainly for debug.